requirements.txt
anthropic>=0.39.0
feedparser>=6.0.11
PyYAML>=6.0.2
tzdata>=2024.1
python-dotenv>=1.0.0